requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
python-docx>=0.8.11
numpy>=1.22.0
SQLAlchemy>=2.0.0
PyCryptodome>=3.14.0